Removed puppet and all git submodules
Puppet is hard to manage and slow, we'll use a PHP solution like we always do.
Git submodules make a developers life much harder, especially if the developers mess around with their git repositories. This is the reason why I'm moving Open Sans back to bower. The git repositories master branch was removed and only the GitHub pages branch is left. This always displays errors and is problematic.