Finished feature/logging
Created real flexible and PSR-3 compatible static log system. We use a static system because singletons are bad. Still, because the class is final and all methods are static testing is pretty complicated. This should be revisited at a later stage and thorough tests should be written. Also possible extension to the current implementation include NewRelic and SMS.