PHP only and implemented using stream wrapper URIs
There are no shell calls anymore necessary for this command plus the usage of URIs makes it totally robust and secure. Note that the execution as root or via sudo is no problem when using our URIs because the stream wrappers make sure that the correct user and group is applied to any file.